fix doc; move some code around
