+++ src/graphnode.ts
@@ -1,10 +1,16 @@
+import { InMemoryCache } from 'apollo-cache-inmemory'
 import { ApolloClient, ApolloQueryResult } from 'apollo-client'
 import { FetchResult, Observable as ZenObservable } from 'apollo-link'
+import { split } from 'apollo-link'
+import { HttpLink } from 'apollo-link-http'
+import { WebSocketLink } from 'apollo-link-ws'
+import { getMainDefinition } from 'apollo-utilities'
 import gql from 'graphql-tag'
+import fetch from 'isomorphic-fetch'
 import { Observable, Observer } from 'rxjs'
 import { catchError, filter, first, map } from 'rxjs/operators'
 import { Logger } from './logger'
-import { createApolloClient, zenToRxjsObservable } from './utils'
+import { zenToRxjsObservable } from './utils'
 
 export interface IApolloQueryOptions {
   fetchPolicy?: 'cache-first' | 'cache-and-network' | 'network-only' | 'cache-only' | 'no-cache' | 'standby',
@@ -14,6 +20,66 @@ export interface IApolloQueryOptions {
 export interface IObservable<T> extends Observable<T> {
   first: () => T
 }
+
+export function createApolloClient(options: {
+  graphqlHttpProvider: string
+  graphqlWsProvider: string
+}) {
+  const httpLink = new HttpLink({
+    credentials: 'same-origin',
+    fetch,
+    uri: options.graphqlHttpProvider
+  })
+
+  const wsLink = new WebSocketLink({
+    options: {
+      reconnect: true
+    },
+    uri: options.graphqlWsProvider,
+    webSocketImpl: WebSocket
+  })
+
+  const wsOrHttpLink = split(
+    // split based on operation type
+    ({ query }) => {
+      const definition = getMainDefinition(query)
+      return definition.kind === 'OperationDefinition' && definition.operation === 'subscription'
+    },
+    wsLink,
+    httpLink
+  )
+  // we can also add error handling
+  // const errorHandlingLink = apollolink.from([
+  //     onerror(({ graphqlerrors, networkerror }) => {
+  //       if (graphqlerrors) {
+  //         graphqlerrors.map(({ message, locations, path }) =>
+  //           console.log(
+  //             `[graphql error]: message: ${message}, location: ${locations}, path: ${path}`,
+  //           ),
+  //         );
+  //       if (networkerror) { console.log(`[network error]: ${networkerror}`)}
+  //       }
+  //     }),
+  //     wsorhttplink
+  //   ])
+  const client = new ApolloClient({
+    cache: new InMemoryCache({
+      cacheRedirects: {
+        Query: {
+          dao: (_, args, { getCacheKey }) =>  getCacheKey({ __typename: 'DAO', id: args.id }),
+          proposal: (_, args, { getCacheKey }) => {
+            // console.log('cache key: ', [args, getCacheKey({ __typename: 'Proposal', id: args.id })])
+            return getCacheKey({ __typename: 'Proposal', id: args.id })
+          }
+        }
+      }
+    }),
+    connectToDevTools: true,
+    link: wsOrHttpLink
+  })
+  return client
+}
+
 /**
  * handles connections with the Graph
  * @param options [description]
